COMMERCE-7217 Moved shop by diagram api module
