+++ modules/apps/commerce/commerce-shop-by-diagram-api/bnd.bnd
@@ -9,4 +9,4 @@ Export-Package:\
 	com.liferay.commerce.shop.by.diagram.service,\
 	com.liferay.commerce.shop.by.diagram.service.persistence,\
 	com.liferay.commerce.shop.by.diagram.type
-Liferay-Enterprise-App: dxp.only=true;product.id=9a473157-06a6-44b6-b017-a360ffaf5f38
+Liferay-Enterprise-App: product.id=9a473157-06a6-44b6-b017-a360ffaf5f38
