+++ modules/apps/commerce/commerce-shop-by-diagram-api/build.gradle
