+++ modules/apps/commerce/commerce-shop-by-diagram-api/src/main/java/com/liferay/commerce/shop/by/diagram/type/CSDiagramType.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.shop.by.diagram.type;
+
+import aQute.bnd.annotation.ProviderType;
+
+import com.liferay.commerce.shop.by.diagram.model.CSDiagramSetting;
+
+import java.util.Locale;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+@ProviderType
+public interface CSDiagramType {
+
+	public String getKey();
+
+	public String getLabel(Locale locale);
+
+	public void render(
+			CSDiagramSetting csDiagramSetting,
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse)
+		throws Exception;
+
+}
