--- modules/dxp/apps/commerce/commerce-shop-by-diagram/commerce-shop-by-diagram-api/src/main/java/com/liferay/commerce/shop/by/diagram/constants/CSDiagramCPTypeConstants.java
@@ -1,24 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- *
- *
- *
- */
-
-package com.liferay.commerce.shop.by.diagram.constants;
-
-/**
- * @author Alessio Antonio Rendina
- */
-public class CSDiagramCPTypeConstants {
-
-	public static final String NAME = "diagram";
-
-}
