--- modules/dxp/apps/commerce/commerce-shop-by-diagram/commerce-shop-by-diagram-api/src/main/java/com/liferay/commerce/shop/by/diagram/constants/CSDiagramWebKeys.java
@@ -1,25 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * The contents of this file are subject to the terms of the Liferay Enterprise
- * Subscription License ("License"). You may not use this file except in
- * compliance with the License. You can obtain a copy of the License by
- * contacting Liferay, Inc. See the License for the specific language governing
- * permissions and limitations under the License, including but not limited to
- * distribution rights of the Software.
- *
- *
- *
- */
-
-package com.liferay.commerce.shop.by.diagram.constants;
-
-/**
- * @author Andrea Sbarra
- */
-public class CSDiagramWebKeys {
-
-	public static final String CS_DIAGRAM_CP_TYPE_DISPLAY_CONTEXT =
-		"CS_DIAGRAM_CP_TYPE_DISPLAY_CONTEXT";
-
-}
